# openapi.yaml
openapi: 3.1.0
info:
  title: Hooter Backend API
  version: '1.0.0'
  description: Internal API for Hooter backend
servers:
  - url: https://hooter_staging.h0oter.com
    description: Staging server

  - url: http://localhost:8888
    description: Local development server
...
paths:
...
  /catalog/categories/top:
    get:
      tags:
        - Catalog
      summary: Get top level categories
      description: Returns all level 0 (top level) Shopify taxonomy categories.
      security:
        - cookieAuth: []

      responses:
        "200":
          description: Top level categories fetched successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - level0
                properties:
                  level0:
                    type: array
                    items:
                      type: object
                      required:
                        - full_name
                        - id
                        - name
                        - vertical
                      properties:
                        full_name:
                          type: string
                          example: Apparel & Accessories
                        id:
                          type: string
                          example: gid://shopify/TaxonomyCategory/aa
                        name:
                          type: string
                          example: Apparel & Accessories
                        vertical:
                          type: integer
                          example: 1

              examples:
                success:
                  summary: Successful request
                  value:
                    level0:
                      - full_name: Animals & Pet Supplies
                        id: gid://shopify/TaxonomyCategory/ap
                        name: Animals & Pet Supplies
                        vertical: 0
                      - full_name: Apparel & Accessories
                        id: gid://shopify/TaxonomyCategory/aa
                        name: Apparel & Accessories
                        vertical: 1
                      - full_name: Electronics
                        id: gid://shopify/TaxonomyCategory/el
                        name: Electronics
                        vertical: 7
                      - full_name: Home & Garden
                        id: gid://shopify/TaxonomyCategory/hg
                        name: Home & Garden
                        vertical: 13
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
        "500":
          description: Could not complete the request
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                  - message
                properties:
                  error:
                    type: string
                    example: failed
                  message:
                    type: string
                    example: could not complete the request
              examples:
                failed:
                  value:
                    error: failed
                    message: could not complete the request
        "default":
          $ref: "#/components/responses/default"
  /catalog/categories/next/{id}:
    get:
      tags:
        - Catalog
      summary: Get next level categories
      description: Returns the child categories of the selected taxonomy category.

      security:
        - cookieAuth: []

      parameters:
        - name: id
          in: path
          required: true
          description: Shopify taxonomy category ID.
          schema:
            type: string
            example: gid://shopify/TaxonomyCategory/ap

        - name: vertical
          in: query
          required: true
          description: Vertical index returned by the top categories endpoint.
          schema:
            type: integer
            example: 0

      responses:
        "200":
          description: Successfully returned the next level categories.
          content:
            application/json:
              schema:
                type: object
                required:
                  - next
                properties:
                  next:
                    type: array
                    items:
                      type: object
                      required:
                        - full_name
                        - id
                        - level
                        - name
                      properties:
                        full_name:
                          type: string
                          example: Apparel & Accessories > Clothing

                        id:
                          type: string
                          example: gid://shopify/TaxonomyCategory/aa-1

                        level:
                          type: integer
                          example: 1

                        name:
                          type: string
                          example: Clothing

              example:
                next:
                  - full_name: Apparel & Accessories > Clothing
                    id: gid://shopify/TaxonomyCategory/aa-1
                    level: 1
                    name: Clothing

                  - full_name: Apparel & Accessories > Clothing Accessories
                    id: gid://shopify/TaxonomyCategory/aa-2
                    level: 1
                    name: Clothing Accessories

                  - full_name: Apparel & Accessories > Costumes & Accessories
                    id: gid://shopify/TaxonomyCategory/aa-3
                    level: 1
                    name: Costumes & Accessories
        "400":
          description: Missing or invalid query parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                    example: vertical(index) and category id is not provided
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
        "default":
          $ref: "#/components/responses/default"
  /catalog/categories/attributes/{id}:
    get:
      tags:
        - Catalog

      summary: Get category attributes

      description: Returns the listing, category, and image attributes for the specified taxonomy category.

      security:
        - cookieAuth: []

      parameters:
        - name: id
          in: path
          required: true
          description: Shopify taxonomy category ID.
          schema:
            type: string
            example: gid://shopify/TaxonomyCategory/aa-1-23-2-1

        - name: vertical
          in: query
          required: true
          description: Vertical index returned by the category endpoints.
          schema:
            type: integer
            example: 1

      responses:
        "200":
          description: Category attributes returned successfully.
          content:
            application/json:
              schema:
                type: object
                required:
                  - listing_attributes
                  - category_attributes
                  - image_attributes

                properties:
                  listing_attributes:
                    type: array
                    items:
                      $ref: "#/components/schemas/AttributeField"

                  category_attributes:
                    type: array
                    items:
                      $ref: "#/components/schemas/AttributeField"

                  image_attributes:
                    type: array
                    items:
                      $ref: "#/components/schemas/ImageAttribute"

              example:
                listing_attributes:
                  - field: product_title
                    handle: product-title
                    name: Product Title
                    description: The title or name of the product.
                    type: string
                    max_length: 200
                    required: true

                  - field: price
                    handle: price
                    name: Price
                    description: The selling price of the product.
                    type: decimal
                    required: true

                category_attributes:
                  - field: product_description
                    handle: product-description
                    name: Product Description
                    description: A detailed description of the product.
                    type: string
                    max_length: 5000
                    required: true

                  - field: fabric
                    handle: fabric
                    name: Fabric
                    description: The primary fabric or material used.
                    type: string
                    max_length: 100
                    required: true

                image_attributes:
                  - field: front
                    handle: front
                    name: Front
                    description: Front view image of the product.
                    order: 0
                    required: true

                  - field: back
                    handle: back
                    name: Back
                    description: Back view image of the product.
                    order: 1
                    required: false
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
        "default":
          $ref: "#/components/responses/default"
  /catalog/categories/bulk-excel-sheet/{id}:
    get:
      tags:
        - Catalog

      summary: Download bulk upload Excel template

      description: |
        Generates and downloads an XLSX template for bulk product uploads.
        The worksheet contains all listing and category-specific fields for the selected taxonomy category.

      security:
        - cookieAuth: []

      parameters:
        - name: id
          in: path
          required: true
          description: Shopify taxonomy category ID.
          schema:
            type: string
            example: gid://shopify/TaxonomyCategory/aa-1-23-2-1

        - name: vertical
          in: query
          required: true
          description: Vertical index returned by the category endpoints.
          schema:
            type: integer
            example: 1

      responses:
        "200":
          description: Excel template generated successfully.
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
        "default":
          $ref: "#/components/responses/default"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    AttributeField:
      type: object
      required:
        - description
        - field
        - handle
        - name
        - required
        - type
      properties:
        description:
          type: string
        field:
          type: string
        handle:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - string
            - integer
            - decimal
            - boolean
        max_length:
          type: integer
          description: Maximum allowed length for string fields.
        required:
          type: boolean
    ImageAttribute:
      type: object
      required:
        - description
        - field
        - handle
        - name
        - order
        - required
      properties:
        description:
          type: string
        field:
          type: string
        handle:
          type: string
        name:
          type: string
        order:
          type: integer
          description: Display order of the image field.
        required:
          type: boolean
...
    loginRequried:
      type: object
      properties:
        status:
          type: string
          example: restricted
        message:
          type: string
          example: "user is not logged in"
    brandRequired:
      type: object
      properties:
        status:
          type: string
          example: restricted
        message:
          type: string
          example: no brand found for this user
...
  responses:
    login_required: # 401
        description: Login session required
        content: 
          application/json:
            schema:
              $ref : "#/components/schemas/loginRequried"
    brand_required: # 403
        description: Brand session required
        content: 
          application/json:
            schema:
              $ref : "#/components/schemas/brandRequired"
...
    "default":
      description: Default response
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - message
            properties:
              status:
                type: string
              message:
                type: string
...